refactor: controller에서 이벤트 제외 로직을 service로 이동
