+++ src/controllers/WebController.js
@@ -1,17 +1,14 @@
-import LottoMachine from '../domains/LottoMachine.js';
 import OutputView from '../views/web/OutputView.js';
 import { getById } from '../views/web/utils/dom.js';
 import LottoResultModal from '../views/web/components/LottoResultModal.js';
 import InputView from '../views/web/InputView.js';
-import WinningService from '../services/WinningService.js';
+import LottoService from '../services/LottoService.js';
 
 class WebController {
-  #lottos;
-  #winningService;
+  #lottoService;
 
   constructor() {
-    this.#lottos = [];
-    this.#winningService = new WinningService();
+    this.#lottoService = new LottoService();
   }
 
   start() {
@@ -37,7 +34,7 @@ class WebController {
     }
     if (target.closest('.resultButton')) {
       event.preventDefault();
-      this.#handleLottoResult();
+      this.#lottoService.handleLottoResult();
       return;
     }
     if (target.closest('.resetButton')) {
@@ -48,42 +45,14 @@ class WebController {
   #addSubmitEvent(event) {
     if (event.target.closest('section.purchase form')) {
       event.preventDefault();
-      this.#handlePurchase();
-    }
-  }
-
-  #handlePurchase() {
-    try {
-      const lottoCount = InputView.enterPurchasePrice();
-      const lottoMachine = new LottoMachine(lottoCount);
-      this.#lottos = lottoMachine.lottos;
-
-      OutputView.printPurchaseLottos(lottoCount, lottoMachine.lottos);
-    } catch (error) {
-      alert(error.message);
-      InputView.resetPurchaseInput();
-    }
-  }
-
-  #handleLottoResult() {
-    try {
-      const { winningNumbers, bonusNumber } = InputView.enterWinningAndBonusNumber();
-      const { winningCounts, profitRate } = this.#winningService.getWinningResult({
-        lottos: this.#lottos,
-        winningNumbers,
-        bonusNumber,
-      });
-
-      OutputView.showModal(winningCounts, profitRate);
-    } catch (error) {
-      alert(error.message);
+      this.#lottoService.handlePurchase();
     }
   }
 
   #handleReset() {
     InputView.resetPurchaseInput();
     OutputView.resetLottoUI();
-    this.#lottos = [];
+    this.#lottoService.resetLotto();
   }
 }
