+++ src/views/web/styles/style.css
@@ -144,7 +144,6 @@
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
   overflow: hidden;
   position: relative;
-  position: relative;
   z-index: 2;
   width: 90%;
   max-width: 400px;
